==> 02-preprocesado.ipynb <==
import os
import json
import gzip
import pandas as pd
import numpy as np
from urllib.request import urlopen
! pip install -r requirements.txt

## 2. Preprocesado

En esta etapa, se realiza el preprocesado de las reseñas para adecuar el formato de las mismas a uno más adecuado para su procesamiento posterior.

### Limpieza de texto

En primer lugar, se realiza una limpieza básica de las reseñas para eliminar caracteres no alfabéticos y convertir el texto a minúsculas.

# Cargamos el dataset preprocesado
reduced_filtered_data = pd.read_csv('reduced_filtered_data.csv')
reduced_filtered_data['reviewText'] = reduced_filtered_data['reviewText'].astype(str)


import re
import string

def clean_text(text):
    # Remove non-alphabetic characters
    text = re.sub(r'[^a-zA-Z]', ' ', text)

    # Convert to lowercase
    text = text.lower()

    return text

# Copiamos en un nuevo dataframe
data_lower = reduced_filtered_data.copy()

# Aplicamos la función de limpieza a los reviews
data_lower['reviewText'] = data_lower['reviewText'].apply(clean_text)

# Guardamos este dataset en un fichero csv
data_lower.to_csv('data_lower.csv', index=False)


# Cargamos el dataset preprocesado
data_lower = pd.read_csv('data_lower.csv')
data_lower['reviewText'] = data_lower['reviewText'].astype(str)
data_lower.shape

### Tokenización
A continuación, se tokenizan las reseñas utilizando la función word_tokenize de la librería nltk.

import nltk
nltk.download('punkt')

from nltk.tokenize import word_tokenize

def tokenize_text(text):
    # Tokenizamos usando word_tokenize
    tokens = word_tokenize(text)

    return tokens

# Aplicamos la función a la columna 'review' del dataset
data_lower['reviewText'] = data_lower['reviewText'].apply(tokenize_text)

### Eliminación de stopwords

Finalmente, se eliminan las palabras vacías (stopwords) de las reseñas utilizando la lista de stopwords de la librería nltk.

nltk.download('stopwords')

from nltk.corpus import stopwords

def remove_stopwords(tokens):
    # removemos stopword de la lista de tokens
    stopwords_list = stopwords.words('english')
    filtered_tokens = [token for token in tokens if token not in stopwords_list]

    return filtered_tokens

# Aplicamos remove_stopwords a la columna review del dataset
data_lower['reviewText'] = data_lower['reviewText'].apply(remove_stopwords)

# Guardamos este dataset en un fichero csv
data_lower.to_csv('data_lower_preprocesado.csv', index=False)
